+++ examples/printer.rs
@@ -1,4 +1,4 @@
-use autorust_openapi::Spec;
+use autorust_openapi::OpenAPI;
 use std::{fs::File, io::Read, process::exit};
 
 pub type Error = Box<dyn std::error::Error>;
@@ -16,7 +16,7 @@ fn main() -> Result<()> {
             // https://github.com/serde-rs/json/issues/160
             let mut bytes = Vec::new();
             File::open(path)?.read_to_end(&mut bytes)?;
-            let spec: Spec = serde_json::from_slice(&bytes)?;
+            let spec: OpenAPI = serde_json::from_slice(&bytes)?;
             println!("# of paths: {}", spec.paths.len());
             for (path, _op) in spec.paths {
                 println!("  {}", path);