+++ src/lib.rs
@@ -1,5 +1,7 @@
+mod openapi;
 mod reference;
 pub mod schema;
 
+pub use self::openapi::*;
 pub use self::reference::*;
 pub use self::schema::*;